Revert "fixed issues"
This reverts commit 05561e2be64ed2f63b77c3c60b389fec78e2a958.

+++ VBN Editor/MDL0Bones.cs
@@ -102,7 +102,7 @@ namespace VBN_Editor
                     int parentid = 0x0FFFFFFF;
                     if (d.pos() != data + 12)
                         parentid = d.readInt();
-                    n.parentIndex = parentid;
+                    n.parentIndex = (uint)parentid;
 
                     n.boneName = d.readString(nameOff, -1).ToCharArray();
                     n.boneId = boneID;
